refactor(r8): extract types/transform/schema/constants from vehicle route (Phase 1)
Pure mechanical extraction — zero behavior change. The R8 vehicle route
(`constructeurs.$brand.$model.$type.tsx`) was 2020 lines; Phase 1 splits
reusable, pure modules into `components/vehicle/r8/`:

- `r8.types.ts`     — 9 interfaces (VehicleData, LoaderData, R8Content, ...)
- `r8-transform.ts` — `transformRpcToLoaderData` (RPC payload → LoaderData)
- `r8-schema.ts`    — `generateVehicleSchema` (Car + BreadcrumbList JSON-LD)
- `r8-constants.ts` — `FAMILY_MICRO_DESCRIPTIONS`
- `index.ts`        — barrel export

Route file: 2020 → 1559 lines (−23%). All exports preserved (loader, meta,
shouldRevalidate, default, ErrorBoundary, handle, links) so the
`$type[.]html.tsx` re-export and `$type[.].tsx` 301-redirect keep working.

Gates: `npx tsc --noEmit` (0 errors), `npx eslint` (0 warnings).

Untouched on purpose (critical SEO):
- 301/410/503 status codes + canonicalization logic in loader
- TecDoc ≥100K remap redirect
- noindex range 60000-83456
- loaderCache (kept in route — will move in Phase 3)

Next: Phase 2 will split the 1180-line `VehicleDetailPage` component into
sub-components per `data-section="S_*"` (S_HERO, S_CATALOG, S_FAQ, ...).
