+++ frontend/app/components/vehicle/r8/index.ts
@@ -1,0 +1,16 @@
+// 🚗 R8 Vehicle Page — Barrel
+// Rôle SEO : R8 - VEHICLE (sélection pièces pour un véhicule spécifique)
+
+export type {
+  VehicleData,
+  CatalogFamily,
+  CatalogGamme,
+  PopularPart,
+  SEOData,
+  R8Block,
+  R8Content,
+  LoaderData,
+} from "./r8.types";
+export { FAMILY_MICRO_DESCRIPTIONS } from "./r8-constants";
+export { generateVehicleSchema } from "./r8-schema";
+export { transformRpcToLoaderData } from "./r8-transform";
