+++ frontend/app/components/vehicle/r8/r8-constants.ts
@@ -1,0 +1,17 @@
+// 🚗 R8 Vehicle Page — Constants
+// Extracted from routes/constructeurs.$brand.$model.$type.tsx (Phase 1 refactor)
+
+export const FAMILY_MICRO_DESCRIPTIONS: Record<string, string> = {
+  Freinage: "Plaquettes, disques, etriers : securite & freinage",
+  Filtration: "Huile, air, carburant, habitacle : entretien moteur",
+  Distribution: "Courroie, galets, pompe a eau : fiabilite",
+  Suspension: "Amortisseurs, ressorts, rotules : tenue de route",
+  Moteur: "Pieces moteur, joints, composants internes",
+  Echappement: "Silencieux, catalyseurs, flexibles",
+  Embrayage: "Kit embrayage, volant moteur, butee",
+  Electrique: "Demarreur, alternateur, batterie : demarrage & charge",
+  Refroidissement: "Radiateur, thermostat, pompe a eau",
+  Direction: "Cremaillere, biellettes, rotules de direction",
+  Carrosserie: "Retroviseurs, pare-chocs, optiques",
+  Essuyage: "Balais, moteur, gicleurs d'essuie-glaces",
+};
