+++ frontend/app/routes/constructeurs.$brand.$model.$type.tsx
@@ -46,16 +46,18 @@ import { logger } from "~/utils/logger";
 import { PageRole, createPageRoleMeta } from "~/utils/page-role.types";
 import { ErrorGeneric } from "../components/errors";
 import { HeroSelection } from "../components/heroes";
-import {
-  ModelContentV1Display,
-  type ModelContentV1Data,
-} from "../components/model";
+import { ModelContentV1Display } from "../components/model";
 import { HtmlContent } from "../components/seo/HtmlContent";
+import {
+  FAMILY_MICRO_DESCRIPTIONS,
+  generateVehicleSchema,
+  transformRpcToLoaderData,
+  type LoaderData,
+} from "../components/vehicle/r8";
 import { hierarchyApi } from "../services/api/hierarchy.api";
 import { brandColorsService } from "../services/brand-colors.service";
 import { isValidImagePath } from "../utils/image-optimizer";
 import { detectMalformedSegment } from "../utils/pieces-route.utils";
-import { stripHtmlForMeta } from "../utils/seo-clean.utils";
 import { normalizeTypeAlias } from "../utils/url-builder.utils";
 
 /**
@@ -75,130 +77,11 @@ export const links: LinksFunction = () => [
 const loaderCache = new Map<string, { data: LoaderData; timestamp: number }>();
 const CACHE_TTL = 60000; // 1 minute
 
-// 📝 Types de données (structure PHP + API /full)
-interface VehicleData {
-  marque_id: number;
-  marque_alias: string;
-  marque_name: string;
-  marque_name_meta: string;
-  marque_name_meta_title: string;
-  marque_logo: string;
-  marque_relfollow: number;
-  modele_id: number;
-  modele_alias: string;
-  modele_name: string;
-  modele_name_meta: string;
-  modele_relfollow: number;
-  modele_pic?: string;
-  type_id: number;
-  type_alias: string;
-  type_name: string;
-  type_name_meta: string;
-  type_power_ps: string;
-  type_body: string;
-  type_fuel: string;
-  type_month_from: string;
-  type_year_from: string;
-  type_month_to: string | null;
-  type_year_to: string | null;
-  type_relfollow: number;
-  power: string;
-  date: string;
-  // 🔧 Codes moteur (depuis API /full)
-  motor_codes?: string[];
-  motor_codes_formatted?: string;
-  // 🔧 Types mines / CNIT (depuis API /full)
-  mine_codes?: string[];
-  mine_codes_formatted?: string;
-  cnit_codes?: string[];
-  cnit_codes_formatted?: string;
-  // 📊 Données techniques formatées
-  power_formatted?: string;
-  cylinder_cm3?: number;
-  production_date_formatted?: string;
-}
-
-interface CatalogFamily {
-  mf_id: number;
-  mf_name: string;
-  mf_description: string;
-  mf_pic: string;
-  gammes: CatalogGamme[];
-}
-
-interface CatalogGamme {
-  pg_id: number;
-  pg_alias: string;
-  pg_name: string;
-}
-
-interface PopularPart {
-  cgc_pg_id: number;
-  pg_alias: string;
-  pg_name: string;
-  pg_name_meta: string;
-  pg_img: string;
-  addon_content: string;
-}
-
-interface _MetaTagsAriane {
-  mta_id: number;
-  mta_title: string;
-  mta_descrip: string;
-  mta_keywords: string;
-  mta_ariane: string;
-  mta_h1: string;
-  mta_content: string;
-  mta_alias: string;
-  mta_relfollow: number;
-}
-
-interface SEOData {
-  title: string;
-  description: string;
-  keywords: string;
-  h1: string;
-  content: string;
-  content2: string;
-  robots: string;
-  canonical: string;
-}
-
-// R8 enriched content (from __seo_r8_pages, optional overlay)
-interface R8Block {
-  id: string;
-  type: string;
-  title: string;
-  renderedText: string;
-  specificityWeight: number;
-}
-
-interface R8Content {
-  h1: string;
-  metaTitle: string;
-  metaDescription: string;
-  blocks: R8Block[];
-  seoDecision: string;
-  diversityScore: number;
-}
-
-interface LoaderData {
-  vehicle: VehicleData;
-  catalogFamilies: CatalogFamily[];
-  popularParts: PopularPart[];
-  seo: SEOData;
-  breadcrumb: {
-    items?: Array<{ name: string; url: string }>;
-    // Legacy support
-    brand: string;
-    model: string;
-    type: string;
-  };
-  // V1 Content - Encyclopedic content for model pages (optional)
-  modelContentV1?: ModelContentV1Data | null;
-  // R8 enriched content (from R8VehicleEnricherService, optional)
-  r8Content?: R8Content | null;
-}
+// Types, transform, schema, constants moved to components/vehicle/r8/
+// - r8.types.ts        : VehicleData, CatalogFamily, PopularPart, SEOData, R8Block, R8Content, LoaderData
+// - r8-transform.ts    : transformRpcToLoaderData (RPC → LoaderData, pure)
+// - r8-schema.ts       : generateVehicleSchema (Car + BreadcrumbList JSON-LD)
+// - r8-constants.ts    : FAMILY_MICRO_DESCRIPTIONS
 
 // ⚡ Contrôle de revalidation pour éviter les rechargements inutiles
 export function shouldRevalidate({
@@ -214,240 +97,6 @@ export function shouldRevalidate({
 
   // Ne recharger que si les paramètres de route changent
   return currentUrl.pathname !== nextUrl.pathname;
-}
-
-// ========================================
-// 🚀 RPC OPTIMISÉ - Transformation des données
-// ========================================
-
-/**
- * Transforme la réponse RPC en LoaderData
- * Compatible avec la structure existante de la page
- */
-function transformRpcToLoaderData(
-  rpcData: any,
-  params: { brand: string; model: string; type: string },
-): LoaderData {
-  const v = rpcData.vehicle;
-  const seoCustom = rpcData.seo_custom;
-
-  // Parsing des paramètres URL (pour fallback)
-  const brandParts = params.brand.split("-");
-  const marque_id = parseInt(brandParts[brandParts.length - 1]) || 0;
-  const marque_alias = brandParts.slice(0, -1).join("-");
-
-  const modelParts = params.model.split("-");
-  const modele_id = parseInt(modelParts[modelParts.length - 1]) || 0;
-  const modele_alias = modelParts.slice(0, -1).join("-");
-
-  const typeWithoutHtml = params.type.replace(".html", "");
-  const typeParts = typeWithoutHtml.split("-");
-  const type_id = parseInt(typeParts[typeParts.length - 1]) || 0;
-
-  // Formatage de la date (logique PHP)
-  let type_date = "";
-  if (!v.type_year_to) {
-    type_date = `du ${v.type_month_from}/${v.type_year_from}`;
-  } else {
-    type_date = `de ${v.type_year_from} à ${v.type_year_to}`;
-  }
-
-  // Données véhicule
-  const vehicleData: VehicleData = {
-    marque_id: v.marque_id || marque_id,
-    marque_alias: v.marque_alias || marque_alias,
-    marque_name: v.marque_name,
-    marque_name_meta: v.marque_name_meta || v.marque_name,
-    marque_name_meta_title: v.marque_name_meta_title || v.marque_name,
-    marque_logo: v.marque_logo || `${v.marque_alias}.webp`,
-    marque_relfollow: v.marque_relfollow || 1,
-    modele_id: v.modele_id || modele_id,
-    modele_alias: v.modele_alias || modele_alias,
-    modele_name: v.modele_name,
-    modele_name_meta: v.modele_name_meta || v.modele_name,
-    modele_relfollow: v.modele_relfollow || 1,
-    modele_pic: v.modele_pic,
-    type_id: v.type_id || type_id,
-    type_alias: normalizeTypeAlias(v.type_alias, v.type_name),
-    type_name: v.type_name,
-    type_name_meta: v.type_name_meta || v.type_name,
-    type_power_ps: v.type_power_ps,
-    type_body: v.type_body,
-    type_fuel: v.type_fuel,
-    type_month_from: v.type_month_from,
-    type_year_from: v.type_year_from,
-    type_month_to: v.type_month_to,
-    type_year_to: v.type_year_to,
-    type_relfollow: v.type_relfollow || 1,
-    power: v.type_power_ps,
-    date: type_date,
-    motor_codes: rpcData.motor_codes || [],
-    motor_codes_formatted: (rpcData.motor_codes || []).join(", "),
-    mine_codes: rpcData.mine_codes || [],
-    mine_codes_formatted: (rpcData.mine_codes || []).join(", "),
-    cnit_codes: rpcData.cnit_codes || [],
-    cnit_codes_formatted: (rpcData.cnit_codes || []).join(", "),
-    power_formatted: v.type_power_ps ? `${v.type_power_ps} ch` : "",
-    cylinder_cm3: v.type_liter
-      ? Math.round(parseFloat(v.type_liter) * 1000)
-      : undefined,
-    production_date_formatted: type_date,
-  };
-
-  // Système SEO avec switch dynamique
-  const getSeoSwitch = (alias: number, typeId: number): string => {
-    const switches: Record<number, string[]> = {
-      1: ["à prix discount", "pas cher", "à mini prix", "en promotion"],
-      2: ["et équipements", "et accessoires", "neuves", "d'origine"],
-      10: [
-        "Toutes les pièces auto",
-        "Trouvez toutes les pièces",
-        "Catalogue complet",
-        "Pièces détachées",
-      ],
-      11: [
-        "Toutes les références",
-        "L'ensemble des pièces",
-        "Toutes les gammes",
-        "Tous les produits",
-      ],
-      12: [
-        "nos fournisseurs certifiés",
-        "nos partenaires agréés",
-        "nos distributeurs",
-        "nos fournisseurs",
-      ],
-    };
-    const options = switches[alias] || [""];
-    return options[typeId % options.length];
-  };
-
-  // SEO (priorité aux données personnalisées)
-  let seoTitle: string;
-  let seoDescription: string;
-  let seoKeywords: string;
-  let h1: string;
-  let content: string;
-  let content2: string;
-
-  if (seoCustom) {
-    seoTitle =
-      seoCustom.mta_title ||
-      `Pièces ${vehicleData.marque_name_meta_title} ${vehicleData.modele_name_meta} ${vehicleData.type_name_meta}`;
-    seoDescription = seoCustom.mta_descrip || "";
-    seoKeywords =
-      seoCustom.mta_keywords ||
-      `${vehicleData.marque_name_meta}, ${vehicleData.modele_name_meta}, ${vehicleData.type_name_meta}`;
-    h1 =
-      seoCustom.mta_h1 ||
-      `${vehicleData.marque_name} ${vehicleData.modele_name} ${vehicleData.type_name} ${vehicleData.type_power_ps} ch ${type_date}`;
-    content = seoCustom.mta_content || "";
-    content2 = "";
-  } else {
-    const comp_switch_title = getSeoSwitch(1, type_id);
-    const comp_switch_desc = getSeoSwitch(2, type_id);
-    seoTitle = `Pièces ${vehicleData.marque_name_meta_title} ${vehicleData.modele_name_meta} ${vehicleData.type_name_meta} ${comp_switch_title}`;
-    seoDescription = `Catalogue pièces détachées pour ${vehicleData.marque_name_meta} ${vehicleData.modele_name_meta} ${vehicleData.type_name_meta} ${vehicleData.type_power_ps} ch ${type_date} neuves ${comp_switch_desc}`;
-    seoKeywords = `${vehicleData.marque_name_meta}, ${vehicleData.modele_name_meta}, ${vehicleData.type_name_meta}, ${vehicleData.type_power_ps} ch, ${type_date}`;
-    h1 = `${vehicleData.marque_name} ${vehicleData.modele_name} ${vehicleData.type_name} ${vehicleData.type_power_ps} ch ${type_date}`;
-    content = `Accédez au catalogue de pièces compatibles avec votre <b>${vehicleData.marque_name} ${vehicleData.modele_name} ${vehicleData.type_name}</b> ${vehicleData.type_power_ps} ch (${vehicleData.type_year_from}–${vehicleData.type_year_to || "aujourd'hui"}). Les gammes ci-dessous sont filtrées selon la motorisation, la période de production et la carrosserie. Pour confirmer la compatibilité, utilisez votre VIN / CNIT (carte grise) : cela évite les erreurs entre versions proches.`;
-    content2 = `En cas de doute, notre équipe valide la référence avant expédition.${vehicleData.motor_codes_formatted ? ` Code moteur connu : <strong>${vehicleData.motor_codes_formatted}</strong>.` : ""}`;
-  }
-
-  // Canonical URL (uses normalizeTypeAlias to handle null, empty, or "null" string values)
-  const canonicalLink = `https://www.automecanik.com/constructeurs/${vehicleData.marque_alias}-${vehicleData.marque_id}/${vehicleData.modele_alias}-${vehicleData.modele_id}/${vehicleData.type_alias}-${vehicleData.type_id}.html`;
-
-  // Catalogue (depuis RPC)
-  const catalogFamilies: CatalogFamily[] = (
-    rpcData.catalog?.families || []
-  ).map((f: any) => ({
-    mf_id: parseInt(f.mf_id),
-    mf_name: f.mf_name,
-    mf_description: f.mf_description || `Système ${f.mf_name.toLowerCase()}`,
-    mf_pic: f.mf_pic || `${f.mf_name.toLowerCase()}.webp`,
-    gammes: (f.gammes || []).map((g: any) => ({
-      pg_id: g.pg_id,
-      pg_alias: g.pg_alias,
-      pg_name: g.pg_name,
-    })),
-  }));
-
-  // Pièces populaires (depuis RPC)
-  const generateSeoContent = (
-    pgName: string,
-    vd: VehicleData,
-    tid: number,
-  ): string => {
-    const switches = ["Achetez", "Trouvez", "Commandez", "Choisissez"];
-    const qualities = ["d'origine", "de qualité", "certifiées", "garanties"];
-    return `${switches[tid % switches.length]} ${pgName} ${vd.marque_name_meta} ${vd.modele_name_meta} ${vd.type_name_meta}, ${qualities[(tid + 1) % qualities.length]} à prix bas.`;
-  };
-
-  const popularParts: PopularPart[] = (rpcData.popular_parts || []).map(
-    (p: any, idx: number) => ({
-      cgc_pg_id: p.pg_id,
-      pg_alias: p.pg_alias,
-      pg_name: p.pg_name,
-      pg_name_meta: p.pg_name_meta || p.pg_name.toLowerCase(),
-      pg_img: p.pg_img || null,
-      addon_content: generateSeoContent(p.pg_name, vehicleData, type_id + idx),
-    }),
-  );
-
-  // Validation SEO pour robots
-  const seoValidation = rpcData.seo_validation || {};
-  let pageRobots = "index, follow";
-  if (!seoValidation.is_indexable) {
-    pageRobots = "noindex, nofollow";
-  }
-
-  // Nouveaux véhicules TecDoc (60000-83456) : noindex jusqu'à validation SEO
-  if (type_id >= 60000 && type_id <= 83456) {
-    pageRobots = "noindex, nofollow";
-  }
-
-  // Blog content - RPC returns simple bsm_* fields, but ModelContentV1Data expects full structure
-  // For now, set to null. To get full model content, a separate API call would be needed.
-  const modelContentV1: ModelContentV1Data | null = null;
-
-  return {
-    vehicle: vehicleData,
-    catalogFamilies,
-    popularParts,
-    seo: {
-      title: seoTitle,
-      description: stripHtmlForMeta(seoDescription),
-      keywords: seoKeywords,
-      h1,
-      content,
-      content2,
-      robots: pageRobots,
-      canonical: canonicalLink,
-    },
-    breadcrumb: {
-      items: [
-        { name: "Accueil", url: "/" },
-        { name: "Constructeurs", url: "/constructeurs" },
-        {
-          name: vehicleData.marque_name,
-          url: `/constructeurs/${vehicleData.marque_alias}-${vehicleData.marque_id}.html`,
-        },
-        {
-          name: vehicleData.modele_name,
-          url: "",
-        },
-        {
-          name: `${vehicleData.type_name} ${vehicleData.type_power_ps} ch`,
-          url: "",
-        },
-      ],
-      brand: vehicleData.marque_name,
-      model: vehicleData.modele_name,
-      type: vehicleData.type_name,
-    },
-    modelContentV1,
-    r8Content: rpcData.r8_content || null,
-  };
 }
 
 // 🚀 Loader optimisé avec RPC (remplace 4 appels API → 1 seul)
@@ -648,101 +297,6 @@ export async function loader({ params }: LoaderFunctionArgs) {
   return defer(loaderData as unknown as Record<string, unknown>);
 }
 
-// 🚗 Générer le breadcrumb structuré Schema.org
-// 🚗 Génère le schema @graph complet: Car + BreadcrumbList
-function generateVehicleSchema(
-  vehicle: VehicleData,
-  breadcrumb: LoaderData["breadcrumb"],
-) {
-  const baseUrl = "https://www.automecanik.com";
-  const canonicalUrl = `${baseUrl}/constructeurs/${vehicle.marque_alias}-${vehicle.marque_id}/${vehicle.modele_alias}-${vehicle.modele_id}/${vehicle.type_alias}-${vehicle.type_id}.html`;
-
-  return {
-    "@context": "https://schema.org",
-    "@graph": [
-      // 1️⃣ Car - Véhicule complet avec toutes les specs
-      {
-        "@type": "Car",
-        "@id": `${canonicalUrl}#vehicle`,
-        name: `${vehicle.marque_name} ${vehicle.modele_name} ${vehicle.type_name}`,
-        brand: { "@type": "Brand", name: vehicle.marque_name },
-        manufacturer: { "@type": "Organization", name: vehicle.marque_name },
-        model: vehicle.modele_name,
-        vehicleConfiguration: vehicle.type_name,
-        // 📅 Année modèle
-        ...(vehicle.type_year_from && {
-          vehicleModelDate: vehicle.type_year_from,
-        }),
-        // 🔧 Moteur
-        vehicleEngine: {
-          "@type": "EngineSpecification",
-          name: vehicle.type_name,
-          ...(vehicle.type_power_ps && {
-            enginePower: {
-              "@type": "QuantitativeValue",
-              value: parseInt(vehicle.type_power_ps),
-              unitCode: "HP",
-            },
-          }),
-        },
-        // ⛽ Carburant
-        ...(vehicle.type_fuel && { fuelType: vehicle.type_fuel }),
-        // 🚗 Carrosserie
-        ...(vehicle.type_body && { bodyType: vehicle.type_body }),
-        // 📅 Période de production
-        ...(vehicle.type_year_from && {
-          additionalProperty: [
-            {
-              "@type": "PropertyValue",
-              name: "Période de production",
-              value: vehicle.type_year_to
-                ? `${vehicle.type_year_from}-${vehicle.type_year_to}`
-                : `depuis ${vehicle.type_year_from}`,
-            },
-          ],
-        }),
-        url: canonicalUrl,
-      },
-      // 2️⃣ BreadcrumbList - Fil d'ariane
-      {
-        "@type": "BreadcrumbList",
-        "@id": `${canonicalUrl}#breadcrumb`,
-        itemListElement: [
-          {
-            "@type": "ListItem",
-            position: 1,
-            name: "Accueil",
-            item: `${baseUrl}/`,
-          },
-          {
-            "@type": "ListItem",
-            position: 2,
-            name: "Constructeurs",
-            item: `${baseUrl}/constructeurs`,
-          },
-          {
-            "@type": "ListItem",
-            position: 3,
-            name: breadcrumb.brand,
-            item: `${baseUrl}/constructeurs/${vehicle.marque_alias}-${vehicle.marque_id}.html`,
-          },
-          {
-            "@type": "ListItem",
-            position: 4,
-            name: breadcrumb.model,
-            item: `${baseUrl}/constructeurs/${vehicle.marque_alias}-${vehicle.marque_id}/${vehicle.modele_alias}-${vehicle.modele_id}.html`,
-          },
-          {
-            "@type": "ListItem",
-            position: 5,
-            name: `${breadcrumb.type} ${vehicle.type_power_ps} ch`,
-          },
-        ],
-      },
-    ],
-  };
-}
-
 // 🎯 Meta function avec SEO optimisé (logique PHP)
 export const meta: MetaFunction<typeof loader> = ({ data: rawData }) => {
   const data = rawData as LoaderData | undefined;
@@ -796,21 +350,6 @@ export const meta: MetaFunction<typeof loader> = ({ data: rawData }) => {
   }
 
   return result;
-};
-
-const FAMILY_MICRO_DESCRIPTIONS: Record<string, string> = {
-  Freinage: "Plaquettes, disques, etriers : securite & freinage",
-  Filtration: "Huile, air, carburant, habitacle : entretien moteur",
-  Distribution: "Courroie, galets, pompe a eau : fiabilite",
-  Suspension: "Amortisseurs, ressorts, rotules : tenue de route",
-  Moteur: "Pieces moteur, joints, composants internes",
-  Echappement: "Silencieux, catalyseurs, flexibles",
-  Embrayage: "Kit embrayage, volant moteur, butee",
-  Electrique: "Demarreur, alternateur, batterie : demarrage & charge",
-  Refroidissement: "Radiateur, thermostat, pompe a eau",
-  Direction: "Cremaillere, biellettes, rotules de direction",
-  Carrosserie: "Retroviseurs, pare-chocs, optiques",
-  Essuyage: "Balais, moteur, gicleurs d'essuie-glaces",
 };
 
 // 🎨 Composant principal avec logique PHP intégrée
